+++ tests/test_all_bathymetry_into_dem_wellington_with_roughness/test_all_bathymetry_into_dem_wellington_with_roughness.py
@@ -115,13 +115,15 @@ class ProcessorRiverBathymetryTest(unittest.TestCase):
             "doesn't exist"
         )
 
-        for file in cls.results_dir.glob("*"):  # only files
-            if file.is_file():
-                file.unlink()
-            elif file.is_dir():
-                shutil.rmtree(file)
-        if cls.results_dir.exists():
-            shutil.rmtree(cls.results_dir)
+        # Cycle through all folders within the cache dir deleting their contents
+        for path in cls.cache_dir.iterdir():
+            if path.is_dir():
+                for file in path.glob("*"):  # only files
+                    if file.is_file():
+                        file.unlink()
+                    elif file.is_dir():
+                        shutil.rmtree(file)
+                shutil.rmtree(path)
 
     @pytest.mark.skipif(sys.platform != "win32", reason="Windows test - this is strict")
     def test_result_geofabric_windows(self):
