+++ tests/test_processor_remote_tiles_wellington_ground_only/test_processor_remote_tiles_wellington_ground_only.py
@@ -117,13 +117,15 @@ class ProcessorRemoteTilesWellingtonTest(unittest.TestCase):
             "doesn't exist"
         )
 
-        for file in cls.results_dir.glob("*"):  # only files
-            if file.is_file():
-                file.unlink()
-            elif file.is_dir():
-                shutil.rmtree(file)
-        if cls.results_dir.exists():
-            shutil.rmtree(cls.results_dir)
+        # Cycle through all folders within the cache dir deleting their contents
+        for path in cls.cache_dir.iterdir():
+            if path.is_dir():
+                for file in path.glob("*"):  # only files
+                    if file.is_file():
+                        file.unlink()
+                    elif file.is_dir():
+                        shutil.rmtree(file)
+                shutil.rmtree(path)
 
     def test_correct_dataset(self):
         """A test to see if the correct dataset is downloaded"""
